Inizia WishlistController

+++ src/main/java/org/generation/italy/reshare/model/repositories/abstractions/AppUserRepository.java
@@ -3,6 +3,6 @@ package org.generation.italy.reshare.model.repositories.abstractions;
 import org.generation.italy.reshare.model.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
+public interface AppUserRepository extends JpaRepository<AppUser, Long> {
     AppUser findByEmail(String email);
 }
